Fix scenario where db_hash = None on close
If the NAME was already patched, `load_remote_db` would never set
self.db_hash. This caused the comparison to always fail during `close`
and the DB to unnecessarily be uploaded.

This also introduces more logging using the `django_s3_sqlite` namespace
to make it easier to turn on/off for debugging.

Some helper functions were added to make the code more succinct while
also avoiding ever reading the file into memory more than once for a
given method.

Since we already have it, the MD5 sum is pushed to S3 to ensure the
integrity of the upload.